+++ api/src/main/java/net/jqwik/api/arbitraries/StringArbitrary.java
@@ -14,16 +14,23 @@ public interface StringArbitrary extends Arbitrary<String> {
 
 	/**
 	 * Set the maximum allowed length {@code maxLength} (included) of generated strings.
+	 *
+	 * @throws IllegalArgumentException if maxLength < 0 or maxLength < min length
 	 */
 	StringArbitrary ofMaxLength(int maxLength);
 
 	/**
-	 * Set the minimum allowed length {@code mixLength} (included) of generated strings.
+	 * Set the minimum allowed length {@code minLength} (included) of generated strings.
+	 * This will also set the max length of the string if {@code minLength} is larger than the current max length.
+	 *
+	 * @throws IllegalArgumentException if minLength < 0
 	 */
 	StringArbitrary ofMinLength(int minLength);
 
 	/**
 	 * Fix the length to {@code length} of generated strings.
+	 *
+	 * @throws IllegalArgumentException if length < 0
 	 */
 	default StringArbitrary ofLength(int length) {
 		return ofMinLength(length).ofMaxLength(length);