+++ engine/src/main/java/net/jqwik/engine/properties/arbitraries/DefaultStringArbitrary.java
@@ -84,13 +84,25 @@ public class DefaultStringArbitrary extends TypedCloneable implements StringArbi
 
 	@Override
 	public StringArbitrary ofMinLength(int minLength) {
+		if (minLength < 0) {
+			throw new IllegalArgumentException("minLength must be between 0 and 2147483647");
+		}
 		DefaultStringArbitrary clone = typedClone();
 		clone.minLength = minLength;
+		clone.maxLength = Math.max(maxLength, minLength);
 		return clone;
 	}
 
 	@Override
 	public StringArbitrary ofMaxLength(int maxLength) {
+		if (maxLength < 0) {
+			throw new IllegalArgumentException("maxLength must be between 0 and 2147483647");
+		}
+		if (maxLength < minLength) {
+			String message = String.format("minLength (%s) must not be larger than maxLength (%s)", minLength, maxLength);
+			throw new IllegalArgumentException(message);
+		}
+
 		DefaultStringArbitrary clone = typedClone();
 		clone.maxLength = maxLength;
 		return clone;